Invert CRC seed in mkey tests
The expected CRC values are calculated according to the actual behavior
of the DV sig API. There was a bug in the API implementation. So, the
MKEY test follows the buggy behavior.

The bug is fixed now in the commit "mlx5: Fix inverted CRC seed for
block signature". So let's fix the tests as well.

Signed-off-by: Sergey Gorenko <sergeygo@nvidia.com>